Fix rubocop Linters manually
