+++ option.rb
@@ -1,6 +1,8 @@
 require './app'
 
 class Options
+  attr_accessor :app
+
   def initialize
     @app = App.new
   end
@@ -19,10 +21,6 @@ class Options
       @app.create_a_rental
     when '6'
       @app.list_rentals
-    when '7'
-      @app.save_people
-      @app.save_books
-      @app.save_rentals
     end
   end
 end
