+++ people.json
@@ -1,1 +1,1 @@
-[{"type":"Student","name":"momo","id":1,"age":"8","parent_permission":true},{"type":"Teacher","name":"mohsen","id":2,"age":"29","specialization":"math teacher"},{"type":"Student","name":"omar","id":3,"age":"7","parent_permission":false},{"type":"Teacher","name":"mohsen","id":4,"age":"mohsen","specialization":"math"},{"type":"Teacher","name":"nasra","id":5,"age":"50","specialization":"history teacher"}]
+[{"type":"Student","name":"momo","id":1,"age":"8","parent_permission":true},{"type":"Teacher","name":"mohsen","id":2,"age":"29","specialization":"math teacher"},{"type":"Student","name":"omar","id":3,"age":"7","parent_permission":false},{"type":"Teacher","name":"mohsen","id":4,"age":"mohsen","specialization":"math"},{"type":"Teacher","name":"nasra","id":5,"age":"50","specialization":"history teacher"},{"type":"Student","name":"ahmed","id":6,"age":"6","parent_permission":true},{"type":"Teacher","name":"hazem","id":7,"age":"29","specialization":"arabic teacher"},{"type":"Teacher","name":"hassan","id":8,"age":"40","specialization":"french teacher"},{"type":"Student","name":"hossam mostafa","id":9,"age":"14","parent_permission":true}]
